renamed components folder to src. made blur filter adjustable
