+++ src/brushes/BrushTemplate.js
