+++ src/brushes/defaultBrush.js
