+++ src/brushes/eraserBrush.js
