+++ src/brushes/sketchyBrush.js
