+++ src/external-lib/gif.js
