+++ src/filters/areacolor.js
