+++ src/filters/edgedetection.js
