+++ src/filters/grayscale.js
