+++ src/filters/invert.js
