+++ src/utils/BrushManager.js
